refactor: move jwk to its own package
To avoid cyclic dependency with did:key creation from kms keys, the jwk api needs to be in its own subpackage.

closes #2921
