+++ pkg/didcomm/protocol/didexchange/keys.go
@@ -14,7 +14,8 @@ import (
 
 	"github.com/hyperledger/aries-framework-go/pkg/crypto"
 	"github.com/hyperledger/aries-framework-go/pkg/doc/did"
-	"github.com/hyperledger/aries-framework-go/pkg/doc/jose"
+	"github.com/hyperledger/aries-framework-go/pkg/doc/jose/jwk"
+	"github.com/hyperledger/aries-framework-go/pkg/doc/jose/jwk/jwksupport"
 	"github.com/hyperledger/aries-framework-go/pkg/kms"
 )
 
@@ -49,12 +50,12 @@ func createSigningVM(km kms.KeyManager, vmType string, keyType kms.KeyType) (*di
 	case ed25519VerificationKey2018, bls12381G2Key2020:
 		return did.NewVerificationMethodFromBytes(vmID, vmType, "", pubKeyBytes), nil
 	case jsonWebKey2020:
-		jwk, err := jose.PubKeyBytesToJWK(pubKeyBytes, keyType)
+		j, err := jwksupport.PubKeyBytesToJWK(pubKeyBytes, keyType)
 		if err != nil {
 			return nil, fmt.Errorf("failed to convert public key to JWK for VM: %w", err)
 		}
 
-		return did.NewVerificationMethodFromJWK(vmID, vmType, "", jwk)
+		return did.NewVerificationMethodFromJWK(vmID, vmType, "", j)
 	default:
 		return nil, fmt.Errorf("unsupported verification method: '%s'", vmType)
 	}
@@ -79,12 +80,12 @@ func createEncryptionVM(km kms.KeyManager, vmType string, keyType kms.KeyType) (
 
 		return did.NewVerificationMethodFromBytes(vmID, vmType, "", key.X), nil
 	case jsonWebKey2020:
-		jwk, err := buildJWKFromBytes(kaPubKeyBytes)
+		j, err := buildJWKFromBytes(kaPubKeyBytes)
 		if err != nil {
 			return nil, fmt.Errorf("createEncryptionVM: %w", err)
 		}
 
-		vm, err := did.NewVerificationMethodFromJWK(vmID, vmType, "", jwk)
+		vm, err := did.NewVerificationMethodFromJWK(vmID, vmType, "", j)
 		if err != nil {
 			return nil, fmt.Errorf("createEncryptionVM: %w", err)
 		}
@@ -95,7 +96,7 @@ func createEncryptionVM(km kms.KeyManager, vmType string, keyType kms.KeyType) (
 	}
 }
 
-func buildJWKFromBytes(pubKeyBytes []byte) (*jose.JWK, error) {
+func buildJWKFromBytes(pubKeyBytes []byte) (*jwk.JWK, error) {
 	pubKey := &crypto.PublicKey{}
 
 	err := json.Unmarshal(pubKeyBytes, pubKey)
@@ -103,7 +104,7 @@ func buildJWKFromBytes(pubKeyBytes []byte) (*jose.JWK, error) {
 		return nil, fmt.Errorf("failed to unmarshal JWK for KeyAgreement: %w", err)
 	}
 
-	var jwk *jose.JWK
+	var j *jwk.JWK
 
 	switch pubKey.Type {
 	case "EC":
@@ -112,7 +113,7 @@ func buildJWKFromBytes(pubKeyBytes []byte) (*jose.JWK, error) {
 			return nil, err
 		}
 
-		jwk = &jose.JWK{
+		j = &jwk.JWK{
 			JSONWebKey: gojose.JSONWebKey{
 				Key:   ecKey,
 				KeyID: pubKey.KID,
@@ -121,7 +122,7 @@ func buildJWKFromBytes(pubKeyBytes []byte) (*jose.JWK, error) {
 			Crv: pubKey.Curve,
 		}
 	case "OKP":
-		jwk = &jose.JWK{
+		j = &jwk.JWK{
 			JSONWebKey: gojose.JSONWebKey{
 				Key:   pubKey.X,
 				KeyID: pubKey.KID,
@@ -131,7 +132,7 @@ func buildJWKFromBytes(pubKeyBytes []byte) (*jose.JWK, error) {
 		}
 	}
 
-	return jwk, nil
+	return j, nil
 }
 
 // nolint:gochecknoglobals
