+++ pkg/doc/signature/suite/bbsblssignature2020/public_key_verifier_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/hyperledger/aries-framework-go/pkg/doc/jose"
+	"github.com/hyperledger/aries-framework-go/pkg/doc/jose/jwk"
 	sigverifier "github.com/hyperledger/aries-framework-go/pkg/doc/signature/verifier"
 )
 
@@ -49,7 +49,7 @@ message2
 	// Failed as we do not support JWK for Bls12381G2Key2020.
 	err = verifier.Verify(&sigverifier.PublicKey{
 		Type: "Bls12381G2Key2020",
-		JWK: &jose.JWK{
+		JWK: &jwk.JWK{
 			Kty: "EC",
 			Crv: "BLS12381_G2",
 		},
