+++ pkg/doc/verifiable/example_credential_test.go
@@ -15,7 +15,7 @@ import (
 	"time"
 
 	"github.com/hyperledger/aries-framework-go/pkg/common/log"
-	"github.com/hyperledger/aries-framework-go/pkg/doc/jose"
+	"github.com/hyperledger/aries-framework-go/pkg/doc/jose/jwk/jwksupport"
 	"github.com/hyperledger/aries-framework-go/pkg/doc/signature/jsonld"
 	"github.com/hyperledger/aries-framework-go/pkg/doc/signature/suite"
 	"github.com/hyperledger/aries-framework-go/pkg/doc/signature/suite/bbsblssignature2020"
@@ -430,7 +430,7 @@ func ExampleCredential_AddLinkedDataProofMultiProofs() {
 	ed25519Suite := ed25519signature2018.New(suite.WithVerifier(ed25519signature2018.NewPublicKeyVerifier()))
 	jsonWebSignatureSuite := jsonwebsignature2020.New(suite.WithVerifier(jsonwebsignature2020.NewPublicKeyVerifier()))
 
-	jwk, err := jose.JWKFromKey(ecdsaSigner.PublicKey())
+	j, err := jwksupport.JWKFromKey(ecdsaSigner.PublicKey())
 	if err != nil {
 		panic(err)
 	}
@@ -449,7 +449,7 @@ func ExampleCredential_AddLinkedDataProofMultiProofs() {
 				return &sigverifier.PublicKey{
 					Type:  "JsonWebKey2020",
 					Value: ecdsaSigner.PublicKeyBytes(),
-					JWK:   jwk,
+					JWK:   j,
 				}, nil
 			}
