+++ test/bdd/pkg/sidetree/sidetree.go
@@ -17,7 +17,7 @@ import (
 	"github.com/trustbloc/sidetree-core-go/pkg/versions/1_0/client"
 
 	diddoc "github.com/hyperledger/aries-framework-go/pkg/doc/did"
-	"github.com/hyperledger/aries-framework-go/pkg/doc/jose"
+	"github.com/hyperledger/aries-framework-go/pkg/doc/jose/jwk"
 	"github.com/hyperledger/aries-framework-go/pkg/vdr/fingerprint"
 	"github.com/hyperledger/aries-framework-go/test/bdd/pkg/util"
 )
@@ -58,9 +58,9 @@ type didResolution struct {
 type CreateDIDParams struct {
 	URL             string
 	KeyID           string
-	JWK             *jose.JWK
-	UpdateJWK       *jose.JWK
-	RecoveryJWK     *jose.JWK
+	JWK             *jwk.JWK
+	UpdateJWK       *jwk.JWK
+	RecoveryJWK     *jwk.JWK
 	KeyType         string
 	ServiceEndpoint string
 }
@@ -120,8 +120,8 @@ func getOpaqueDocument(params *CreateDIDParams) ([]byte, error) {
 	return doc.Bytes()
 }
 
-func getPubKey(jwk *jose.JWK) (string, error) {
-	publicKey, err := pubkey.GetPublicKeyJWK(jwk.Key)
+func getPubKey(j *jwk.JWK) (string, error) {
+	publicKey, err := pubkey.GetPublicKeyJWK(j.Key)
 	if err != nil {
 		return "", err
 	}
@@ -134,7 +134,7 @@ func getPubKey(jwk *jose.JWK) (string, error) {
 	return string(opsPubKeyBytes), nil
 }
 
-func getCreateRequest(doc []byte, updateJWK, recoveryJWK *jose.JWK) ([]byte, error) {
+func getCreateRequest(doc []byte, updateJWK, recoveryJWK *jwk.JWK) ([]byte, error) {
 	pubKeyUpdate, err := pubkey.GetPublicKeyJWK(updateJWK.Key)
 	if err != nil {
 		return nil, err
